pyproject.toml
[project]
name = "storysparkaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "langchain==1.0.4",
    "langchain-google-genai==3.0.1",
    "langchain-ollama==1.0.0",
    "langchain-openai>=1.0.2",
    "langsmith>=0.4.41",
    "pydantic==2.12.4",
    "pylatex==1.4.2",
    "python-dotenv==1.2.1",
]
